requirements.txt
setuptools==40.6.3
Flask==1.0.2
Flask_WTF==0.14.2
Werkzeug==0.15.3
python_Levenshtein==0.12.0
apscheduler==3.5.3
bleach==3.3.0
beautifulsoup4==4.7.1
pymongo==3.7.2
cexprtk==0.3.3
chardet==3.0.4
python_dateutil==2.7.5
flask_mail==0.9.1
flask_session==0.3.1
natsort==5.5.0
pandas==1.1.4
Requests==2.21.0
Flask-Babel==0.12.2
PyLTI==0.7.0
flask-talisman==0.7.0
nose2==0.9.1
selenium==3.141.0
bandit==1.6.2
pycryptodome==3.7.3
python3-saml==1.9.0